# baseline/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/homework/submissions/{id}/grade": {
      "post": {
        "operationId": "HomeworkController_grade",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Homework"
        ]
      }
    },
    "/api/homework/submissions/{id}/return": {
      "post": {
        "operationId": "HomeworkController_returnForRedo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Homework"
        ]
      }
    },
    "/api/homework/teachers/{teacherId}/pending-grading": {
      "post": {
        "operationId": "HomeworkController_listPendingByTeacher",
        "parameters": [
          {
            "name": "teacherId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Homework"
        ]
      }
    },
    "/api/homework/students/{studentId}/list": {
      "post": {
        "operationId": "HomeworkController_listByStudent",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Homework"
        ]
      }
    },
...
    "/api/homework/db/submissions/{id}/grade": {
      "post": {
        "operationId": "HomeworkController_gradeInDb",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Homework"
        ]
      }
    },
    "/api/homework/db/submissions/{id}/return": {
      "post": {
        "operationId": "HomeworkController_returnForRedoInDb",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Homework"
        ]
      }
    },
    "/api/homework/db/teachers/{teacherId}/assignments": {
      "post": {
        "operationId": "HomeworkController_listAssignmentsByTeacherInDb",
        "parameters": [
          {
            "name": "teacherId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Homework"
        ]
      }
    },
    "/api/homework/db/students/{studentId}/assignments": {
      "post": {
        "operationId": "HomeworkController_listAssignmentsByStudentInDb",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Homework"
        ]
      }
    },
    "/api/homework/db/teachers/{teacherId}/pending-grading": {
      "post": {
        "operationId": "HomeworkController_listPendingByTeacherInDb",
        "parameters": [
          {
            "name": "teacherId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Homework"
        ]
      }
    },
...
  },
  "info": {
    "title": "Edu Server API",
    "description": "教培 CRM 后端 API — Phase B.L3 contract source-of-truth",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}